requirements.txt
dill==0.3.6
Flask==2.2.3
numpy~=1.21.6
pandas~=1.3.5
PyYAML==6.0
scikit_learn==1.2.2
setuptools==65.6.3
six==1.16.0
streamlit==1.21.0

yaml~=0.2.5
scikit-learn~=1.0.2
pillow~=9.5.0